+++ docs/releases.md
@@ -5,6 +5,27 @@
 - [Github Sponsor page](https://github.com/sponsors/simulot)
 - [paypal donor page](https://www.paypal.com/donate/?hosted_button_id=VGU2SQE88T2T4)
 
+### feat: exclude files based on a pattern
+
+Use the `-exclude-files=PATTERN` to exclude certain files or directories from the upload. Repeat the option for each pattern do you need. The following directories are excluded automatically:
+- @eaDir/
+- @__thumb/
+- SYNOFILE_THUMB_\*.\*
+- Lightroom Catalog/
+- thumbnails/
+- .DS_Store/
+
+Example, the following command excludes any files in directories called backup or draft and any file with name finishing with "copy)" as PXL_20231006_063121958 (another copy).jpg:
+```sh
+immich-go -sever=xxxxx -key=yyyyy upload -exclude-files=backup/ -exclude-files=draft/ -exclude=copy).*  /path/to/your/files
+```
+
+fixes:
+- [#365](https://github.com/simulot/immich-go/issues/365) missing associated metadata file isn't correct
+- [#299](https://github.com/simulot/immich-go/issues/299) Real time GUI log only shows 4 lines #299
+- [#370](https://github.com/simulot/immich-go/issues/370) ui: clearly mention when the upload in completed
+- [#232](https://github.com/simulot/immich-go/issues/232) Exclude based on filename / glob
+
 ## Release 0.19.1
 
 ### fix: UploadAsset